requirements.txt
flask
langchain
openai
pinecone
pypdf
python-dotenv
langchain-community
langchain-pinecone
langchain-openai
langchain-core
langchain-experimental
langchain-community
